Hide framework classes from ROOT in SiStripTools
- separate classes that do the calculation from those that hold
  the results
- Use ROOT dictionaries only for results to avoid causing additional
  auto parsing.
