+++ DPGAnalysis/SiStripTools/plugins/MultiplicityAlgorithms.cc
@@ -1,41 +1,38 @@
-#include "DPGAnalysis/SiStripTools/interface/Multiplicities.h"
+#include "MultiplicityAlgorithms.h"
 
-ClusterSummarySingleMultiplicity::ClusterSummarySingleMultiplicity()
-    : m_subdetenum(ClusterSummary::STRIP), m_varenum(ClusterSummary::NCLUSTERS), m_mult(0), m_collection() {}
+using namespace sistriptools::algorithm;
 
 ClusterSummarySingleMultiplicity::ClusterSummarySingleMultiplicity(const edm::ParameterSet& iConfig,
                                                                    edm::ConsumesCollector&& iC)
     : m_subdetenum((ClusterSummary::CMSTracker)iConfig.getParameter<int>("subDetEnum")),
       m_varenum((ClusterSummary::VariablePlacement)iConfig.getParameter<int>("varEnum")),
-      m_mult(0),
       m_collection(iC.consumes<ClusterSummary>(iConfig.getParameter<edm::InputTag>("clusterSummaryCollection"))) {}
 
 ClusterSummarySingleMultiplicity::ClusterSummarySingleMultiplicity(const edm::ParameterSet& iConfig,
                                                                    edm::ConsumesCollector& iC)
     : m_subdetenum((ClusterSummary::CMSTracker)iConfig.getParameter<int>("subDetEnum")),
       m_varenum((ClusterSummary::VariablePlacement)iConfig.getParameter<int>("varEnum")),
-      m_mult(0),
       m_collection(iC.consumes<ClusterSummary>(iConfig.getParameter<edm::InputTag>("clusterSummaryCollection"))) {}
 
-void ClusterSummarySingleMultiplicity::getEvent(const edm::Event& iEvent, const edm::EventSetup& iSetup) {
-  m_mult = 0;
+ClusterSummarySingleMultiplicity::value_t ClusterSummarySingleMultiplicity::getEvent(
+    const edm::Event& iEvent, const edm::EventSetup& iSetup) const {
+  int mult = 0;
 
   edm::Handle<ClusterSummary> clustsumm;
   iEvent.getByToken(m_collection, clustsumm);
 
   switch (m_varenum) {
     case ClusterSummary::NCLUSTERS:
-      m_mult = int(clustsumm->getNClus(m_subdetenum));
+      mult = int(clustsumm->getNClus(m_subdetenum));
       break;
     case ClusterSummary::CLUSTERSIZE:
-      m_mult = int(clustsumm->getClusSize(m_subdetenum));
+      mult = int(clustsumm->getClusSize(m_subdetenum));
       break;
     case ClusterSummary::CLUSTERCHARGE:
-      m_mult = int(clustsumm->getClusCharge(m_subdetenum));
+      mult = int(clustsumm->getClusCharge(m_subdetenum));
       break;
     default:
-      m_mult = -1;
+      mult = -1;
   }
+  return value_t(mult);
 }
-
-int ClusterSummarySingleMultiplicity::mult() const { return m_mult; }
