+++ DPGAnalysis/SiStripTools/src/classes.h
@@ -8,27 +8,4 @@
 
 #include "DataFormats/Common/interface/Wrapper.h"
 
-namespace DPGAnalysis_SiStripTools {
-  struct dictionary {
-    TinyEventCollection dummycoll;
-    SingleSiStripDigiMultiplicity dummy1;
-    SingleSiPixelClusterMultiplicity dummy2;
-    SingleSiStripClusterMultiplicity dummy3;
-    SiPixelClusterSiStripClusterMultiplicityPair dummy4;
-    ClusterSummarySingleMultiplicity dummy5;
-    ClusterSummaryMultiplicityPair dummy6;
-
-    edm::Wrapper<TinyEventCollection> dummywrappedcoll;
-    edm::Wrapper<EventWithHistory> dummywrappedEWH;
-    edm::Wrapper<APVCyclePhaseCollection> dummywrappedAPVC;
-
-    /*
-    edm::Wrapper<SingleSiStripDigiMultiplicity> dummywrapped1;
-    edm::Wrapper<SingleSiStripClusterMultiplicity> dummywrapped2;
-    edm::Wrapper<SingleSiPixelClusterMultiplicity> dummywrapped3;
-    edm::Wrapper<SiPixelClusterSiStripClusterMultiplicityPair> dummywrapped4;
-    */
-  };
-}  // namespace DPGAnalysis_SiStripTools
-
 #endif  // DPGAnalysis_SiStripTools_classes_H
